pyproject.toml
[tool.poetry]
name = "openllms"
version = "0.1.0"
description = ""
authors = ["matheusfvesco <114014793+matheusfvesco@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.37.1"
python-dotenv = "^1.0.1"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
